Fix for Issue #447 'test_palo_alto_collectors'.py has been updated.
During testing, an issue was discovered and fixed in the 'palo_alto_parsers.py'. Getting an ARP table with a single entry only returns a "dict". We were expecting a list of dictionaries, added a check to wrap a single dictionary in a list to ensure the data format is as expected.